Merged PR 593: Run migrations before app service deployment
This changes the order of the deploy steps to run migrations first and then deploy to App Service second.

We should aim to make migrations backwards compatible in order to do zero downtime deploys. This is following the [expand and contract pattern](https://martinfowler.com/bliki/ParallelChange.html), and whilst it's not important on systems which are not released, it's good to follow best practice to make it easier for ourselves in future.

The latest version of the code will probably not work without the latest schema changes, so deploying it first gives us no benefit. If we make the schema changes backwards compatible between versions we can do the schema migration, and not worry about any delay or problems in the app service deployment.

As an example, if the code change involves replacing one column with another, the way to do this would be:

1. First PR adds new column and changes app to use the new column
2. Followed up by a second PR to remove the old now unused column

In a lot of cases schemas grow rather than shrink in complexity so step 2 isn't necessary!

If there's worries about consistency by having two versions of the system deployed then these can be dealt with as appropriate by:

* Code that checks and syncs both places - and is then tidied up after (this is taking the expand and contract pattern further)
* Running reports to find any discrepancies which slipped through during the upgrade and fixing them manually
* Scheduling downtime

But hopefully in a lot of cases we can avoid these altogether!