Merged PR 1133: Add pre-push git commit hook to show PRs url
Add pre-push git commit hook to show PRs url

This is something I've been wanting for a while. Some platforms (e.g. BitBucket, and I think GitHub) will give you a URL to create a PR after you push, so you don't have to manually navigate to it. Azure doesn't do this, but we can build the url for the PRs page from the git remote, and log it to the console when you push. Annoyingly, there's no post-push hook, so we've got to use pre-push