Merged PR 1714: Remove EF bulk insert because it's more trouble than it's worth
Remove EF bulk insert because it's more trouble than it's worth

I've tried to use this on a couple of projects recently, and I don't really think it's worth it. The reason it's so fast is because it doesn't track objects or set their IDs, and it can't automatically work out some of the relationships that normally work in EF (e.g. if you're using a link table for a many-to-many relationship). With all these issues, it can be quite hard to get the data seeder to seed complicated objects, which can lead to quite a lot of developer frustration. I think it's not worth the pain, given that seeding isn't run very often and we should hopefully not be seeding too much data anyway